+++ mafia-wizard-master/src/main/resources/db/V202109051406__add_game_master_table.sql
@@ -9,7 +9,7 @@ CREATE TABLE wizard.game_masters (
 
 CREATE TABLE wizard.games (
                                      game_uuid      UUID          NOT NULL,
-                                     game_number    int4          not NULL default 0,
+                                     game_number    int8          not NULL default 0,
                                      players        JSON           null,
                                      updated_at            TIMESTAMPTZ   NULL default now(),
                                      created_at            TIMESTAMPTZ   NULL default now(),
